+++ mpi4jax/__init__.py
@@ -10,7 +10,6 @@ from ._src import (  # noqa: E402
     allgather,
     allreduce,
     alltoall,
-    auto_tokenize,
     barrier,
     bcast,
     gather,
