+++ mpi4jax/_src/__init__.py
@@ -20,7 +20,6 @@ from .collective_ops.send import send  # noqa: F401
 from .collective_ops.sendrecv import sendrecv  # noqa: F401
 
 from .utils import has_cuda_support  # noqa: F401
-from .tokenizer import auto_tokenize  # noqa: F401
 
 # check version of jaxlib
 from . import jax_compat
