+++ mpi4jax/_src/collective_ops/allreduce.py
@@ -21,7 +21,6 @@ from ..decorators import translation_rule_cpu, translation_rule_gpu
 from ..validation import enforce_types
 from ..comm import get_default_comm
 from ..jax_compat import Tracer, Token
-from ..tokenizer import token_override_registry
 
 # The Jax primitive
 mpi_allreduce_p = Primitive("allreduce_mpi")  # Create the primitive
@@ -65,14 +64,6 @@ def allreduce(x, op, *, comm=None, token=None):
     op = wrap_as_hashable(op)
     comm = wrap_as_hashable(comm)
     return tuple(mpi_allreduce_p.bind(x, token, op=op, comm=comm, transpose=False))
-
-
-def mpi_allreduce_token_override(in_args, new_token, op, comm, transpose):
-    x, _ = in_args
-    return mpi_allreduce_p.bind(x, new_token, op=op, comm=comm, transpose=transpose)
-
-
-token_override_registry[mpi_allreduce_p] = mpi_allreduce_token_override
 
 
 # This function compiles the operation
@@ -202,8 +193,6 @@ def mpi_allreduce_transpose_rule(tan_args, *x_args, op, comm, transpose):
     )
     return res, token_tan
 
-
-token_override_registry[mpi_allreduce_p] = mpi_allreduce_token_override
 
 mpi_allreduce_p.multiple_results = True
 mpi_allreduce_p.def_impl(mpi_allreduce_impl)
