+++ mpi4jax/_src/collective_ops/bcast.py
@@ -21,7 +21,6 @@ from ..decorators import translation_rule_cpu, translation_rule_gpu
 from ..validation import enforce_types
 from ..comm import get_default_comm
 from ..jax_compat import Tracer, Token
-from ..tokenizer import token_override_registry
 
 # The Jax primitive
 mpi_bcast_p = Primitive("bcast_mpi")  # Create the primitive
@@ -71,14 +70,6 @@ def bcast(x, root, *, comm=None, token=None):
         return (x, token)
 
     return (res, token)
-
-
-def mpi_bcast_token_override(in_args, new_token, root, comm):
-    x, _ = in_args
-    return mpi_bcast_p.bind(x, new_token, root=root, comm=comm)
-
-
-token_override_registry[mpi_bcast_p] = mpi_bcast_token_override
 
 
 # This function compiles the operation
