+++ mpi4jax/_src/collective_ops/reduce.py
@@ -21,7 +21,6 @@ from ..decorators import translation_rule_cpu, translation_rule_gpu
 from ..validation import enforce_types
 from ..comm import get_default_comm
 from ..jax_compat import Tracer, Token
-from ..tokenizer import token_override_registry
 
 # The Jax primitive
 mpi_reduce_p = Primitive("reduce_mpi")  # Create the primitive
@@ -70,14 +69,6 @@ def reduce(x, op, root, *, comm=None, token=None):
         return (x, token)
 
     return (res, token)
-
-
-def mpi_reduce_token_override(in_args, new_token, op, root, comm):
-    x, _ = in_args
-    return mpi_reduce_p.bind(x, new_token, op=op, root=root, comm=comm)
-
-
-token_override_registry[mpi_reduce_p] = mpi_reduce_token_override
 
 
 # This function compiles the operation
