+++ mpi4jax/experimental/__init__.py
@@ -1,0 +1,6 @@
+from . import register_overrides  # noqa: F401
+from .tokenizer import auto_tokenize
+
+__all__ = [
+    "auto_tokenize",
+]
